app/requirements.txt
streamlit==1.29.0
pandas==2.1.0
PyYAML==6.0.1
plotly==5.18.0
python-dotenv==1.0.0
pillow==10.1.0
numpy==1.26.2
